aarch64 cleanup and calls
