+++ polynomials/src/lib.rs
@@ -9,17 +9,22 @@
 //!
 //! FIXME: This is kyber specific for now.
 
-pub use libcrux_traits::{Operations, GenericOperations, FIELD_ELEMENTS_IN_VECTOR, FIELD_MODULUS};
+use libcrux_traits::INVERSE_OF_MODULUS_MOD_MONTGOMERY_R;
+pub use libcrux_traits::{GenericOperations, Operations, FIELD_ELEMENTS_IN_VECTOR, FIELD_MODULUS};
 
 // There's no runtime detection here. This either exposes the real SIMD vector,
 // or the portable when the feature is not set.
 //
 // The consumer needs to use runtime feature detection and the appropriate vector
 // in each case.
+#[cfg(feature = "simd128")]
+pub use libcrux_polynomials_aarch64::SIMD128Vector;
 #[cfg(feature = "simd256")]
 pub use libcrux_polynomials_avx2::SIMD256Vector;
 #[cfg(not(feature = "simd256"))]
 pub type SIMD256Vector = PortableVector;
+#[cfg(not(feature = "simd128"))]
+pub type SIMD128Vector = PortableVector;
 
 /// Values having this type hold a representative 'x' of the Kyber field.
 /// We use 'fe' as a shorthand for this type.
@@ -37,7 +42,6 @@ pub type FieldElementTimesMontgomeryR = i32;
 
 pub(crate) const MONTGOMERY_SHIFT: u8 = 16;
 pub(crate) const MONTGOMERY_R: i32 = 1 << MONTGOMERY_SHIFT;
-pub(crate) const INVERSE_OF_MODULUS_MOD_MONTGOMERY_R: u32 = 62209; // FIELD_MODULUS^{-1} mod MONTGOMERY_R
 
 pub(crate) const BARRETT_SHIFT: i64 = 26;
 pub(crate) const BARRETT_R: i64 = 1 << BARRETT_SHIFT;
